requirements.txt
accelerate==0.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.2
filelock==3.14.0
fsspec==2024.3.1
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psutil==5.9.8
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pyttsx3==2.90
pytweening==1.2.0
pywin32==306
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
safetensors==0.4.3
scipy==1.13.0
setuptools==69.5.1
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
sympy==1.12
tbb==2021.12.0
tesseract==0.1.3
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
typing_extensions==4.11.0
urllib3==2.2.1
win10toast==0.9
